src/tryonservice/requirements.txt
google-generativeai==0.8.5
Pillow
python-dotenv
fastapi
uvicorn
python-multipart
